requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
requests
openai
google-generativeai
python-multipart
re
